requirements.txt
python-telegram-bot[callback-data]
qbittorrent-api
transliterate
wakeonlan